First stab at linux firefox support.
-Mikeal



git-svn-id: http://svn.getwindmill.com/trunk@137 78c7df6f-8922-0410-bcd3-9426b1ad491b

+++ windmill/browser/firefox.py
@@ -145,7 +145,9 @@ class MozillaBrowser(object):
         os.kill(self.p_id, signal)
         try:
             os.kill(self.p_id, 0)
-            os.kill(self.p_id+1, signal)
+            if sys.platform == 'darwin':
+                os.kill(self.p_id+1, signal)
+                os.kill(self.p_id, signal)
         except:
             pass
